pyproject.toml
[tool.poetry]
name = "Nady-System-Back-End"
version = "1.0.0"
description = "Nady Django App using Django & Graphql"
authors = ["Shady El Nady <r>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
graphene-django = "^2.15.0"
python-decouple = "^3.4"
django-graphql-jwt = "^0.3.1"
PyJWT = "^1.7.1"
graphene-django-optimizer = "^0.8.0"
django-filter = "^2.4.0"
django-debug-toolbar = "^3.2.1"
graphene-file-upload = "^1.3.0"
djongo = "1.3.6"
pymongo = "3.12.3"
Pillow = "^9.2.0"
django-polymorphic-tree = "^2.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
pytest = "^6.2.4"
pytest-django = "^4.2.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
